+++ stable_baselines/common/buffers.py
@@ -1,12 +1,14 @@
 import random
+from typing import Optional, List, Union
 
 import numpy as np
 
 from stable_baselines.common.segment_tree import SumSegmentTree, MinSegmentTree
+from stable_baselines.common.vec_env import VecNormalize
 
 
 class ReplayBuffer(object):
-    def __init__(self, size):
+    def __init__(self, size: int):
         """
         Implements a ring buffer (FIFO).
 
@@ -17,7 +19,7 @@ class ReplayBuffer(object):
         self._maxsize = size
         self._next_idx = 0
 
-    def __len__(self):
+    def __len__(self) -> int:
         return len(self._storage)
 
     @property
@@ -26,11 +28,11 @@ class ReplayBuffer(object):
         return self._storage
 
     @property
-    def buffer_size(self):
+    def buffer_size(self) -> int:
         """float: Max capacity of the buffer"""
         return self._maxsize
 
-    def can_sample(self, n_samples):
+    def can_sample(self, n_samples: int) -> bool:
         """
         Check if n_samples samples can be sampled
         from the buffer.
@@ -40,7 +42,7 @@ class ReplayBuffer(object):
         """
         return len(self) >= n_samples
 
-    def is_full(self):
+    def is_full(self) -> int:
         """
         Check whether the replay buffer is full or not.
 
@@ -86,7 +88,27 @@ class ReplayBuffer(object):
                 self._storage[self._next_idx] = data
             self._next_idx = (self._next_idx + 1) % self._maxsize
 
-    def _encode_sample(self, idxes):
+    @staticmethod
+    def _normalize_obs(obs: np.ndarray,
+                       env: Optional[VecNormalize] = None) -> np.ndarray:
+        """
+        Helper for normalizing the observation.
+        """
+        if env is not None:
+            return env.normalize_obs(obs)
+        return obs
+
+    @staticmethod
+    def _normalize_reward(reward: np.ndarray,
+                          env: Optional[VecNormalize] = None) -> np.ndarray:
+        """
+        Helper for normalizing the reward.
+        """
+        if env is not None:
+            return env.normalize_reward(reward)
+        return reward
+
+    def _encode_sample(self, idxes: Union[List[int], np.ndarray], env: Optional[VecNormalize] = None):
         obses_t, actions, rewards, obses_tp1, dones = [], [], [], [], []
         for i in idxes:
             data = self._storage[i]
@@ -96,13 +118,19 @@ class ReplayBuffer(object):
             rewards.append(reward)
             obses_tp1.append(np.array(obs_tp1, copy=False))
             dones.append(done)
-        return np.array(obses_t), np.array(actions), np.array(rewards), np.array(obses_tp1), np.array(dones)
+        return (self._normalize_obs(np.array(obses_t), env),
+                np.array(actions),
+                self._normalize_reward(np.array(rewards), env),
+                self._normalize_obs(np.array(obses_tp1), env),
+                np.array(dones))
 
-    def sample(self, batch_size, **_kwargs):
+    def sample(self, batch_size: int, env: Optional[VecNormalize] = None, **_kwargs):
         """
         Sample a batch of experiences.
 
         :param batch_size: (int) How many transitions to sample.
+        :param env: (Optional[VecNormalize]) associated gym VecEnv
+            to normalize the observations/rewards when sampling
         :return:
             - obs_batch: (np.ndarray) batch of observations
             - act_batch: (numpy float) batch of actions executed given obs_batch
@@ -112,7 +140,7 @@ class ReplayBuffer(object):
                 and 0 otherwise.
         """
         idxes = [random.randint(0, len(self._storage) - 1) for _ in range(batch_size)]
-        return self._encode_sample(idxes)
+        return self._encode_sample(idxes, env=env)
 
 
 class PrioritizedReplayBuffer(ReplayBuffer):
@@ -181,7 +209,7 @@ class PrioritizedReplayBuffer(ReplayBuffer):
         idx = self._it_sum.find_prefixsum_idx(mass)
         return idx
 
-    def sample(self, batch_size, beta=0):
+    def sample(self, batch_size: int, beta: float = 0, env: Optional[VecNormalize] = None):
         """
         Sample a batch of experiences.
 
@@ -191,6 +219,8 @@ class PrioritizedReplayBuffer(ReplayBuffer):
 
         :param batch_size: (int) How many transitions to sample.
         :param beta: (float) To what degree to use importance weights (0 - no corrections, 1 - full correction)
+        :param env: (Optional[VecNormalize]) associated gym VecEnv
+            to normalize the observations/rewards when sampling
         :return:
             - obs_batch: (np.ndarray) batch of observations
             - act_batch: (numpy float) batch of actions executed given obs_batch
@@ -210,7 +240,7 @@ class PrioritizedReplayBuffer(ReplayBuffer):
         max_weight = (p_min * len(self._storage)) ** (-beta)
         p_sample = self._it_sum[idxes] / self._it_sum.sum()
         weights = (p_sample * len(self._storage)) ** (-beta) / max_weight
-        encoded_sample = self._encode_sample(idxes)
+        encoded_sample = self._encode_sample(idxes, env=env)
         return tuple(list(encoded_sample) + [weights, idxes])
 
     def update_priorities(self, idxes, priorities):
@@ -232,4 +262,3 @@ class PrioritizedReplayBuffer(ReplayBuffer):
         self._it_min[idxes] = priorities ** self._alpha
 
         self._max_priority = max(self._max_priority, np.max(priorities))
-