+++ tests/test_vec_normalize.py
@@ -2,7 +2,9 @@ import subprocess
 
 import gym
 import numpy as np
+import pytest
 
+from stable_baselines import DDPG, DQN, SAC, TD3
 from stable_baselines.common.running_mean_std import RunningMeanStd
 from stable_baselines.common.vec_env import (DummyVecEnv, VecNormalize, VecFrameStack,
     sync_envs_normalization, unwrap_vec_normalize)
@@ -123,6 +125,20 @@ def test_normalize_external():
     assert obs.shape == norm_obs.shape
 
 
+@pytest.mark.parametrize("model_class", [DDPG, DQN, SAC, TD3])
+def test_offpolicy_normalization(model_class):
+    if model_class == DQN:
+        env = DummyVecEnv([lambda: gym.make('CartPole-v1')])
+    else:
+        env = DummyVecEnv([make_env])
+    env = VecNormalize(env, norm_obs=True, norm_reward=True, clip_obs=10., clip_reward=10.)
+
+    model = model_class('MlpPolicy', env, verbose=1)
+    model.learn(total_timesteps=1000)
+    # Check getter
+    assert isinstance(model.get_vec_normalize_env(), VecNormalize)
+
+
 def test_sync_vec_normalize():
     env = DummyVecEnv([make_env])
 
@@ -147,6 +163,7 @@ def test_sync_vec_normalize():
 
     obs = env.reset()
     original_obs = env.get_original_obs()
+    dummy_rewards = np.random.rand(10)
     # Normalization must be different
     assert not np.allclose(obs, eval_env.normalize_obs(original_obs))
 
@@ -154,6 +171,7 @@ def test_sync_vec_normalize():
 
     # Now they must be synced
     assert np.allclose(obs, eval_env.normalize_obs(original_obs))
+    assert np.allclose(env.normalize_reward(dummy_rewards), eval_env.normalize_reward(dummy_rewards))
 
 
 def test_mpi_runningmeanstd():